+++ test/wasmex/instance_test.exs
@@ -113,7 +113,7 @@ defmodule Wasmex.InstanceTest do
       receive do
         {:returned_function_call, {:ok, [42]}, :fake_from} -> nil
       after
-        1000 ->
+        2000 ->
           raise "message_expected"
       end
     end